requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django_filters
mysql-connector

